requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.12
django-celery-beat==2.9.0
django-rest-framework==0.1.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
drf-spectacular==0.29.0
gunicorn==25.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycountry==26.2.16
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
